add more tests for DataFrame
